requirements.txt
appnope==0.1.0
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
certifi==2016.2.28
cycler==0.10.0
dask==0.8.1
decorator==4.0.9
functools32==3.2.3.post2
gnureadline==6.3.3
ipython==4.1.2
ipython-genutils==0.1.0
Jinja2==2.8
jsonschema==2.5.1
MarkupSafe==0.23
matplotlib==1.5.1
mpmath==0.19
networkx==1.11
numpy==1.10.4
pandas==0.18.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
Pillow==3.1.1
ptyprocess==0.5.1
pydicom==0.9.9
pyparsing==2.1.0
python-dateutil==2.5.0
pytz==2016.1
pyzmq==15.2.0
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.17.0
seaborn==0.7.0
simplegeneric==0.8.1
SimpleITK==0.8.0
singledispatch==3.4.0.3
six==1.10.0
sympy==1.0
toolz==0.7.4
tornado==4.3
traitlets==4.2.1
